pyproject.toml
[tool.poetry]
name = "comptrain"
version = "0.1.0"
description = ""
authors = ["Alejandro Alcalde <algui91@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
python-telegram-bot = "^12.4.2"
beautifulsoup4 = "^4.8.2"
requests = "^2.23.0"


[tool.poetry.dev-dependencies]
pytest = "^4.6"
black = "^19.10b0"

# [tool.poetry.scripts]
# comptrain = "comptrain.bot:main"
